Add gaming (lutris, steam) packages and nonfreegaming user
Lutris is used because it helps a lot with playing on Wine (especially
when DXVK is not fully supported, it seems), for example Starcraft II.

Steam is used because it's the only way to play some games, and it works
very well on Linux, for example CS:GO or Overcooked 2. Steam also provides
Proton, which helps a lot with playing non-Linux Steam games on Wine, for
example Age of Empires II or Overcooked 1.