Prevent running steam or wine using a regular account
This is a very simple way to avoid running malware or badly coded proprietary
software with enough rights to do actual damage. Running games with a dedicated
account works flawlessly and limits risks to the minimum.

NB: there's a bug in the current release of AwesomeWM which prevents this rc.lua
from working properly. This is addressed in the following PR:

  https://github.com/awesomeWM/awesome/pull/2878