Use the -zz flag of rsync instead of the legacy -z flag
The legacy -z flag is not supported anymore on ArchLinux, following the
removal of the bundled zlib. -zz is pretty much the same thing, except maybe
it compresses a bit less and consumes less CPU as well. The big difference is
that by using the zlib provided by the system, it ensures it is up-to-date and
more secure.

-zz was introduced in rsync 3.1.1 back in 2014.

For more information, see:
  https://www.archlinux.org/news/rsync-compatibility/