Add xsel-connect to xsel-mode to set the $DISPLAY after startup
This is useful for example when an Emacs server is started with systemd
before the user choses to start an X server at all. In this case, the user
can M-x xsel-connect :0 once the X server is running to set the $DISPLAY
properly and copy/paste between the Emacs server and the X server.